+++ src/index.ts
@@ -1,4 +1,7 @@
 export * from './inscription-parser.service';
+export * from './inscription-parser.service.helper';
 export * from './parsed-inscription';
 export * from './decode-data-uri';
 export * from './cbor';
+export * from './src20-parser.service';
+export * from './src20-parser.service.helper';
