+++ src/src20-parser.service.spec.ts
@@ -6,8 +6,8 @@ describe('SRC20 parser', () => {
   it('should parse SRC-20 Transactions', () => {
 
     const txn = readTransaction('50aeb77245a9483a5b077e4e7506c331dc2f628c22046e7d2b4c6ad6c6236ae1');
+    const result = decodeSrc20Transaction(txn);
 
-    const decodedData = decodeSrc20Transaction(txn);
-    console.log(decodedData);
+    expect(result).toEqual('{"p":"src-20","op":"transfer","tick":"STEVE","amt":"100000000"}')
   });
 });
